ref(metrics): Do not set up database unnecessarily
These tests do not use the database and therefore should not use
`sentry.testutils.cases`, but rather `unittest.TestCase`. This improves
runtime on my machine for `pytest ..../test_base_indexer.py` from 10
seconds to half a second.
